requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-ninja==0.19.1
djangorestframework==3.13.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pydantic==1.10.2
python-dateutil==1.5
python-decouple==3.6
python-jose==3.3.0
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
